software/app_ergonomia/requirements.txt
streamlit
requests
python-dotenv
fpdf
matplotlib
